moved authentication module to services/
